backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydicom==2.4.4
numpy==1.26.4
matplotlib==3.10.0
Pillow==10.4.0
python-multipart==0.0.20
requests==2.31.0

